# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ManageHR API",
        "description": "Documentación de la API RESTful de ManageHR desarrollada en Laravel con autenticación JWT",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/hojasvidahasexperiencia": {
            "get": {
                "tags": [
                    "Experiencias en Hojas de Vida"
                ],
                "summary": "Listar todas las experiencias en hojas de vida",
                "operationId": "adaa3645b628cfd4804e69b689ef5486",
                "responses": {
                    "200": {
                        "description": "Listado de relaciones experiencia-hoja de vida"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Experiencias en Hojas de Vida"
                ],
                "summary": "Asociar experiencia a una hoja de vida",
                "operationId": "de339ce88865c94b9ff1928de80fd315",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "idHojaDevida",
                                    "idExperiencia",
                                    "estado"
                                ],
                                "properties": {
                                    "idHojaDevida": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "idExperiencia": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "estado": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "archivo": {
                                        "type": "string",
                                        "example": "ruta/archivo.pdf",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Relación creada correctamente"
                    },
                    "400": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/hojasvidahasexperiencia/{id}": {
            "get": {
                "tags": [
                    "Experiencias en Hojas de Vida"
                ],
                "summary": "Obtener detalle de experiencia en hoja de vida por ID",
                "operationId": "188bc969ae1e3d6fb84325b7043222ee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la relación experiencia-hoja de vida",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle encontrado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Experiencias en Hojas de Vida"
                ],
                "summary": "Actualizar una experiencia laboral en hoja de vida",
                "operationId": "dde78f250bf70e1aede8342828f4ef2a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la experiencia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cargoEmpresa",
                                    "empresa",
                                    "tiempoExperiencia",
                                    "estado"
                                ],
                                "properties": {
                                    "cargoEmpresa": {
                                        "type": "string",
                                        "example": "Desarrollador"
                                    },
                                    "empresa": {
                                        "type": "string",
                                        "example": "Empresa XYZ"
                                    },
                                    "tiempoExperiencia": {
                                        "type": "string",
                                        "example": "1 año"
                                    },
                                    "estado": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Experiencia actualizada correctamente"
                    },
                    "404": {
                        "description": "No se encontró la experiencia"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Experiencias en Hojas de Vida"
                ],
                "summary": "Eliminar relación experiencia-hoja de vida",
                "operationId": "7292f727f339b65f06c9d06b88c5a893",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la relación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eliminado correctamente"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Experiencias en Hojas de Vida"
                ],
                "summary": "Actualizar parcialmente una experiencia laboral",
                "operationId": "242c6e063e3f851a1cf38f13b475c4b4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la experiencia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "cargoEmpresa": {
                                        "type": "string",
                                        "example": "Auxiliar"
                                    },
                                    "empresa": {
                                        "type": "string",
                                        "example": "Empresa ABC"
                                    },
                                    "tiempoExperiencia": {
                                        "type": "string",
                                        "example": "2 años"
                                    },
                                    "estado": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualización parcial exitosa"
                    },
                    "404": {
                        "description": "Experiencia no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/hojasvidahasexperiencia/documento/{numDocumento}": {
            "get": {
                "tags": [
                    "Experiencias en Hojas de Vida"
                ],
                "summary": "Buscar experiencias laborales por número de documento",
                "operationId": "7d29c7dd21baa62d525bd434b4f25333",
                "parameters": [
                    {
                        "name": "numDocumento",
                        "in": "path",
                        "description": "Número de documento del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1234567890
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Experiencias encontradas"
                    },
                    "404": {
                        "description": "Hoja de vida no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/hojasvidahasexperiencia/descargar/{id}": {
            "get": {
                "tags": [
                    "Experiencias en Hojas de Vida"
                ],
                "summary": "Descargar archivo adjunto de experiencia laboral",
                "operationId": "bd3f945dadab58bc171a19e5b9c767bb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del archivo de experiencia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Archivo descargado con éxito",
                        "content": {
                            "application/octet-stream": {}
                        }
                    },
                    "404": {
                        "description": "Archivo no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/hojasvidahasexperiencia/hoja/{idHojaDeVida}": {
            "get": {
                "tags": [
                    "Experiencias en Hojas de Vida"
                ],
                "summary": "Buscar experiencias por hoja de vida",
                "operationId": "2eb213ce5fd8e14d82a2f3b17d6639e2",
                "parameters": [
                    {
                        "name": "idHojaDeVida",
                        "in": "path",
                        "description": "ID de la hoja de vida",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de experiencias encontradas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Introduce tu token JWT en el formato: Bearer {token}"
            }
        }
    },
...
}